requirements.txt
yt-dlp           # 유튜브 영상 정보 및 오디오 추출
faster-whisper   # 음성 인식(STT) 모델 (OpenAI Whisper보다 4-5배 빠름)
google-genai     # Gemini AI API 연동 최신 SDK
python-dotenv    # .env 환경 변수 로드
fastapi          # FastAPI 웹 프레임워크
uvicorn[standard] # ASGI 서버
pydantic         # 데이터 검증
beautifulsoup4   # HTML 파싱 (웹 크롤링)
readability-lxml # 웹 본문 추출
requests         # HTTP 클라이언트


# pip freeze 결과

annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
defusedxml==0.7.1
distro==1.9.0
filelock==3.20.3
fsspec==2026.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-genai==1.59.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
llvmlite==0.46.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
openai-whisper==20250625
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.1
python-dotenv==1.2.1
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
setuptools==80.9.0
sniffio==1.3.1
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
torch==2.9.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
websockets==15.0.1
youtube-transcript-api==0.6.3
yt-dlp==2025.12.8
